personality-backend/requirements.txt
fastapi
uvicorn
pandas
numpy
psycopg2-binary